new: usr: CKEditor support (refs #21)
If the Redmine CKEditor plugin is in use, one button (or two, if the
DMSF plugin is installed) will be added to the toolbar.
Handling of diagrams inside of the editor is not easy because the
document body must be saved before the diagrams, so at this time the
dialogs are only shortcuts for inserting macros.
Maybe I can add embedded editing for already created diagrams.