new: usr: editing support for textile button macros
Placing che cursor in a body macro e clicking the macro button will open
the dialog pre-filled with the values extracted from the macro heading.
This can be useful to change values with a pre-validating dialog instead
of changing he macro directly, with the risk of syntax errors or wrong
parameter names.