fix: usr: fixed SVG corrupted after save before reload (fixes #25)
After editing an SVG diagram and saving it, double clicking on the
diagram without reloading page causes an `Not a diagram` error in the
drawio diagram editor.
Seems that updating the SVG in the page causes the insertion of an
`"=""' text, that confuses the parser.
Removed with an regex, but not identified the source problem (maybe
related to the `XMLSerializer` Firefox browser object).