Fixed generating api key even if api not enabeld
When the api is not enabled there should be no api key requested
for the current user. Moreover, Redmine
MyControllerTest#test_reset_api_key_without_existing_key
would fail since it does not expect the current user to have an api key by
default.

Added a warning for the admin if api is disabled since the plugin needs
the api for saving diagrams as attachment.