Fixes uncaught type error
Math support will only be provided if the user has checked the
corresponding box in plugin settings. When no math support is required
the javascript snippet should be ignored.

Since only MathJax will be ignored and not MathJax.Hub, javascript will
raise a type error. This is fixed with a further condition.